Исправляет коды получения/отправки данных
