+++ js/api.js
@@ -1,3 +1,26 @@
+const ALERT_SHOW_TIME = 5000;
+
+const showAlert = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = 100;
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left ='20%';
+  alertContainer.style.top = '30%';
+  alertContainer.style.right = '20%';
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '30px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+  alertContainer.style.color = 'white';
+
+  alertContainer.textContent = message;
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+};
+
 const getData = (onFail) =>
   fetch('https://23.javascript.pages.academy/keksobooking/data')
     .then((response) => {
@@ -9,6 +32,7 @@ const getData = (onFail) =>
     .then((response) => response.json())
     .catch(() => onFail('Произошла ошибка запроса. Попробуйте ещё раз'));
 
+const offersPromise = getData(showAlert);
 
 const sendData = (onSuccess, onFail, body) => {
   fetch(
@@ -30,4 +54,4 @@ const sendData = (onSuccess, onFail, body) => {
     });
 };
 
-export {getData, sendData};
+export {sendData, offersPromise};
