+++ js/form-validation.js
@@ -1,37 +1,10 @@
-const formAd = document.querySelector('.ad-form');
-const headerFormAd = formAd.querySelector('.ad-form-header');
-const elementsFormAd = formAd.querySelectorAll('.ad-form__element');
-const buttonSubmitFormAd = formAd.querySelector('.ad-form__submit');
-const buttonResetFormAd = formAd.querySelector('.ad-form__reset');
-const formMap = document.querySelector('.map__filters');
-const listFeaturesMap = formMap.querySelector('.map__features');
-const filtersMap = formMap.querySelectorAll('.map__filter');
-
-//ФУНКЦИЯ ПЕРЕКЛЮЧЕНИЯ АКТИВНОГО/НЕАКТИВНОГО СОСТОЯНИЯ
-const  toggleInActiveStatePage  = (isInActive) => {
-  if (isInActive) {
-    formAd.classList.add('ad-form--disabled');
-    formMap.classList.add('map__filters--disabled');
-  }  else {
-    formAd.classList.remove('ad-form--disabled');
-    formMap.classList.remove('map__filters--disabled');
-  }
-  headerFormAd.disabled = isInActive;
-  elementsFormAd.forEach((elementFormAd) => {elementFormAd.disabled = isInActive;});
-  buttonSubmitFormAd.disabled = isInActive;
-  buttonResetFormAd.disabled = isInActive;
-  listFeaturesMap.disabled = isInActive;
-  filtersMap.forEach((filterMap) => {filterMap.disabled = isInActive;});
-};
-
-toggleInActiveStatePage(true);
-
 //ВАЛИДАЦИЯ ФОРМЫ
 const titleAd = document.querySelector('#title');
 const MIN_TITLE_LENGTH = 30;
 const MAX_TITLE_LENGTH = 100;
 
 const roomNumber = document.querySelector('#room_number');
+const roomNumberSelected = roomNumber.querySelector('[selected]');
 const capacityRoom = document.querySelector('#capacity');
 const capacityRoomOptions = capacityRoom.querySelectorAll('option');
 const capacityRoomOptionNoGuests = capacityRoom.querySelector('[value="0"]');
@@ -71,7 +44,7 @@ titleAd.addEventListener('input', () => {
 //Количество комнат и количество мест
 roomNumber.addEventListener('change', (evt) => {
   const selectedRoom = evt.target.value;
-  (capacityRoomOptions).forEach((capacityRoomOption) => {
+  capacityRoomOptions.forEach((capacityRoomOption) => {
     if (selectedRoom < ROOM_NO_GUESTS)  {
       capacityRoomOption.disabled = capacityRoomOption.value > selectedRoom;
       capacityRoomOptionNoGuests.disabled = true;
@@ -106,7 +79,6 @@ typeHousing.addEventListener('change', (evt) => {
   priceAd.placeholder = MinPriceHousing[selectedtypeHousing];
 });
 
-
 //Цена за ночь
 priceAd.addEventListener('input', () => {
   if (priceAd.value > MAX_PRICE) {
@@ -119,4 +91,4 @@ priceAd.addEventListener('input', () => {
   priceAd.reportValidity();
 });
 
-export {toggleInActiveStatePage, formAd, priceAd, buttonResetFormAd};
+export {priceAd, MinPriceHousing, roomNumberSelected, capacityRoomOptions, capacityRoomOptionNoGuests};
