+++ js/main.js
@@ -1,3 +1,3 @@
-import './form.js';
+import './form-validation.js';
+import './form-send.js';
 import './map.js';
-import './send-form.js';
